Use `size_t` for shape values for simplicity

+++ core-cpp/include/inference-engine/InferenceEngine.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <cstdint>
+#include <cstddef>
 #include <vector>
 
 namespace inference_engine
@@ -9,13 +9,13 @@ class InferenceEngine
 {
 public:
     virtual size_t get_input_count() const = 0;
-    virtual const std::vector<int64_t> &get_input_shape(size_t index) const = 0;
-    virtual void set_input_shape(size_t index, const std::vector<int64_t> &shape) = 0;
+    virtual const std::vector<size_t> &get_input_shape(size_t index) const = 0;
+    virtual void set_input_shape(size_t index, const std::vector<size_t> &shape) = 0;
     virtual void set_input_data(size_t index, const float *data) = 0;
 
     virtual size_t get_output_count() const = 0;
-    virtual const std::vector<int64_t> &get_output_shape(size_t index) const = 0;
-    virtual void set_output_shape(size_t index, const std::vector<int64_t> &shape) = 0;
+    virtual const std::vector<size_t> &get_output_shape(size_t index) const = 0;
+    virtual void set_output_shape(size_t index, const std::vector<size_t> &shape) = 0;
     virtual void set_output_data(size_t index, float *data) = 0;
 
     virtual void run() = 0;
